Moved Field, Filter, PersonForm, Person, and PersonsList components (in PhonebookVite project) into their own components to avoid cluttering in App.jsx. Stable to last assignment, 2.15.
