+++ part_2/phonebookVite/src/components/Field.jsx
@@ -1,0 +1,8 @@
+const Field = (props) =>
+{
+	return (
+		<input value={props.value} onChange={props.handler} />
+	)
+}
+
+export default Field
