+++ part_2/phonebookVite/src/components/Person.jsx
@@ -1,0 +1,12 @@
+const Person = ({data, deleteEntry}) =>
+{
+	return (
+		<li>
+			{data.name} &nbsp; 
+			{data.number} &nbsp;
+			<button onClick={deleteEntry}>delete</button>
+		</li>
+	)
+}
+
+export default Person
